:white_check_mark: RootViewModel 테스트 코드 작성
